Experimented with sending Json through sockets.
In order to be able to do more with the project, I decided to send Json
through the socket. The Json can then be decoded to help me build a
dynamic web app. Note: I had to add removeAllListeners() to api.js which
led me to learn that this is a pretty hacky way of achieving my goal. I
will keep it for now but would like to eventually swap the sockets out with
http requests.